fixing type of a hwp variable in command struct

+++ mcp/include/command_struct.h
@@ -671,7 +671,7 @@ struct CommandDataStruct {
     int no_step;
     int use_pot;
     double pot_targ;
-	int margin;
+	float margin;
   } hwpr;
 
   int pin_is_in;
